+++ test/mock/invocation-termination/src/events/event-timeout-async/index.js
@@ -4,12 +4,16 @@ let { join } = require('path')
 exports.handler = async (event) => {
   event = JSON.parse(event.Records[0].Sns.Message)
   let pathToFile = event.path
-  let timeout = 1250 // this event is configured to time out after 1s (see config.arc), so lets set timeout to something slightly higher
-  console.log(`will write to ${pathToFile} in ${timeout}ms...`)
+
+  let timeout = 1250 // this event is configured to time out after 1s (see config.arc), so we'll setTimeout longer
+  console.log(`event-timeout-async will write to ${pathToFile} in ${timeout}ms...`)
   await new Promise(resolve => setTimeout(resolve, timeout))
+
   // this should never execute as timeout is set to 1 sec in config.arc
-  console.log('in timeout, writing to file')
+  console.log(`event-timeout-async writing ${pathToFile}`)
   const writeStream = fs.createWriteStream(pathToFile)
   writeStream.write('hiya')
   writeStream.end()
+  console.log(`event-timeout-async ending now!`)
+  return
 }